# Fix threading issue when reading and writing from SslStream at the same time