@krwq 2019-05-22
`BIO_read`/`BIO_write` was speculation based on their description in the docs (also it would be hard to write reliable test for that) - didn't know we had prevention against that