@krwq 2019-05-22
There seem to be some issue with Azure DevOps - it reports build as red but nothing actually failed on the latest retry: https://dev.azure.com/dnceng/public/_build/results?buildId=191964
Already talked about that with @safern offline and he confirmed the issue